Make sure things are loaded before use
The youtube api now only declares functions when the player is loaded.

Calling playVideo() or pauseVideo() before the player is loaded resulted in a fatal error.

I tried to check if there was an event, but only a global function is supported, which does not scale. So I used our own App.loaded instead.