Fixed Video first resize problem (#42)
* Fixed Video first resize problem

The first resize of the video was done too soon. It was done before it was loaded and had its size. I added a listener for "loadedmetadata" event. When triggered, the first resize is done.

* Update video.js

isVideoResized is not a state of the video. It is a flag to determine if the video is resizable. Changed name to "resizable" to avoid further confusion.