Update preventDefault logic for mobile (#87)
The preventDefault code was executed all the time, even if the min-width and max-width were not respected. This caused some problems with the following use case: on mobile, we want to prevent propagation because the hyperlink is used to toggle a slider. on desktop, we want the hyperlink to act as an hyperlink.