Refactor infinite scroll logic (#89)
The infinite scroll logic was broken. The more content there was, the faster the infinite scroll was triggered.

The reworked logic is based on the bottom of the infinite scroll container relative to the bottom of the window. Let's say that the trigger percentage is 0.8. That means the infinite scroll will load more content when the bottom of the infinite container is 0.8 * windowHeight under the bottom of the window. I think this is the basic behaviour of a infinite scroll: you want to load more content before the bottom of the infinite scroll container is visible.

This logic is more stable than the previous one which frankly, I didn't understand.