Add other lang links detection
The links module is not able to make the framework go to a translated
url. Even worst, we need to update "un-updateable" zone on the screen.

We never ran into that problem, since the alt-lang link normally
suffice. But when client enters url manually, they can point to
alternate versions.