Add support for captchas not in a page or article (#183)
PR #169 removed the ability to pass in a context to load function.
This is now restored.

This commit also adds a container option, to prevent mix up.

In order to resolve items outside the page, resolveTriggersNotInPage()
will compute parents of each triggers and return the proper collection
of context. The module then loads them when recaptcha is
ready.