Add auto native share module (#220)
This module will detect if the window.navigator.share API is available
in the browser. If so, it will hook on .js-native-share and share the
current page. If not, it will remove all .js-native-share from the DOM.

Right now, this API is only supported on Chrome 71+ on android and
Safari 12.2+ on iOS